+++ src/gui/mod.rs
@@ -1,5 +1,6 @@
 pub mod ansi;
 pub mod pathedit;
+pub mod regexedit;
 mod sandbox;
 mod window;
 pub use sandbox::*;